bug #PSBM-49382 aka can't clone VM located inside raid6 dir in pstorage cluster
compiler optimizer re-orders statements execution inside GetDirSizePrivate
so that isDir test is skipped for the config file and nested
GetDirSizePrivate is called, i.e. it treats the file as a folder
which is not.
transformation of code helps to convince the compiler to produce
a code which does what we need.