Disp-to-disp commands for remote restore PSBM-50512
Client sends mountImage command to a backup storage to prepare an NBD device
for remote restoration, later one responds with RestoreImage command - that a
restoration could be started using Url of a prepared NBD device.

Note, rename of the first command is safe, because this command is used for
restoration only and a remote restoration is prohibited now. So, no
compatibility issues could be met - in a local scenario client's and server's
versions are always the same.