Wait until nbd device is ready before returning from Qemu::setImage() (PSBM-51427)
Qemu::setImage() waits 1 second for newly spawned qemu-nbd process to
exit on error. If it doesn't exit, then Qemu::setImage() thinks that the
corresponding nbd device is ready and returns success.
But the nbd device preparation is actually taking place in the background.
The only way to ensure that the device is ready -- is to check its size
using BLKGETSIZE64 ioctl. The size should not be zero.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>