Refactor: move CVmEvent*, BaseNode and ProtoSerializer to prlcommon
The problem is that CVmEvent is a generic class, which is used everywhere in
prl projects. And if one wants to write a common function, which manipulates
CVmEvent - he adds a dependency of libprlcommn on libprlxmlmodel. Actually, it
was already done for ErrorSimple's method convertToEvent().

At the same time - there is no reason to keep this class in xmlmodel - because
its autogeneration is highly customized and the class itself is pretty stable.
So, let's move it to libprlcommon.

There is also a useful side effect of this change - we can finally solve a
duplication of ProtoSerializer in libprlsdk and prl-disp-service. As we have
CVmEvent implemented in libprlcommon now, nothing blocks us from moving of
ProtoSerializer to libprlcommon as well. So, let's do it.