bug #PSBM-56148 aka Slow VM recovery process from one node to another
qemu nbd implemented a new extension to nbd protocol. it improves
performance of a retoration. it is enabled via the -x command line
argument. the url of the server is augmented with value of the
argument. nbd in update3 doesn't support that extension. thus we
need to bump the backup protocol version and use the feature for the
latest one only.