bug #PSBM-66449 aka 11Task_ExecVm was finished with PRL_ERR_OPERATION_FAILED
we introduced a limit on maximum size of a package. the limit is applied
on a receiving side of a socket only. but a sender is still able
to produce huge packages and send them which results in connection abortion.
we need to prohibit construction of that kind of packages.
IOPackage has the setBuffer function to fill a buffer with data. it is
a good place for the limit test. let's force all the execution paths
setting package buffers to impose that test and fail when the limit is
exceeded.