bug #PSBM-96208 aka Unable LIVE restore VM from Remore host
qemu-img performs a test of a backing image for its internal purposes.
we don't need it because the store is always under our control and
the later operation will certainly fail if anything is wrong.
the tool has -u option for create task to avoid touching the image.
let's use it.