bug #PSBM-95242 aka dispatcher failed to create problem-report
we have a limit on a package size to prevent ddos of dispatcher
service. a problem report at the customer side exceedes the
limit and triggered connection abort at a client.
thus we need to not create unlimited packages in dispatcher to
fail early with a proper diagnostic and double the limit to
allow the customer to generate problem reports.