bug #PSBM-96920 aka PlsCtlTimeoutError: Unable to get private ip. Exit code: 1
we use a json parser from boost property_tree both in dispatcher
and in common. the later one is not aware of thread safety which
is a big problem. it is a problem to use the boost library with
new compiler flags. fortunately, we have json-c library installed
with the system. let's use the json parsing feature from the
library in common instead