bug #PSBM-97491 aka bused killed by SIGSEGV
qemu-nbd has issues when 2 instances operate on the same nbd device.
our naive approach for selecting an unallocated nbd device incurs
a race when concurrency is in place cause we don't have any protection
from selecting the same device and use its name as an argument for
qemu-nbd then we have a problem.
let's use a flock on the device inode to prevent using of a device
concurrently. also we need to check if some other user that doesn't
follow the protocol employed the device