switch qemu-nbd handling to async mode
now we cannot catch if qemu-nbd process quits until the user makes
a call of the disk api. this may lead to nbd devices exhaustion
because we track used devices to ensure exclusive access by qemu-nbd
processes. also an unwaited qemu-nbd is weird.
let's handle qemu-nbd lifecycle in async mode and depend on events
from the process and sdk users. the backend needs an event loop
thus a running qt application is required. we use main thread as a
host for the backend. separate 2 types of backend for supported
use cases.