Refactoring: Remove OPENSSL_NO_DEPRECATED macro
According to OpenSSL documentation, if OPENSSL_NO_DEPRECATED macro
is defined, all deprecated public symbols in all OpenSSL versions
up to and including the version given by OPENSSL_API_COMPAT will
be hidden.
So we remove the current macro to get warnings about which parts
of openssl in libprlcommon are out of date.

Signed-off-by: Fima Shevrin <efim.shevrin@virtuozzo.com>